+++ docs/Whats-New.md
@@ -302,6 +302,7 @@ New:
 - `CreateUnit` improvements & additions (units spawning in air, spawn animation) (by Starkku)
 - Option to center pause menu background (by Starkku)
 - Separate power drain and power drain enhancement buildings (by Morton)
+- Reverse rubble (by Morton)
 
 Vanilla fixes:
 - Allow AI to repair structures built from base nodes/trigger action 125/SW delivery in single player missions (by Trsdy)